+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Issue #60](https://github.com/itsallcode/openfasttrace-maven-plugin/issues/55) Import coverage from resource directories
+
 ## [1.7.0] - 2024-02-26
 
 - [PR #57](https://github.com/itsallcode/openfasttrace-maven-plugin/pull/57) Mark plugin as thread safe